+++ jest.config.js
@@ -6,7 +6,7 @@ module.exports = {
     'vue'
   ],
   transform: {
-    '^.+\\.vue$': 'vue-jest',
+    '^.+\\.vue$': '@vue/vue2-jest',
     '.+\\.(css|styl|less|sass|scss|svg|png|jpg|ttf|woff|woff2)$': 'jest-transform-stub',
     '^.+\\.jsx?$': 'babel-jest'
   },
@@ -16,14 +16,13 @@ module.exports = {
   moduleNameMapper: {
     '^@/(.*)$': '<rootDir>/src/$1'
   },
+  setupFilesAfterEnv: ['<rootDir>/tests/unit/setup.js'],
   snapshotSerializers: [
     'jest-serializer-vue'
   ],
+  testEnvironment: 'jsdom',
   testMatch: [
     '**/tests/unit/**/*.spec.(js|jsx|ts|tsx)|**/__tests__/*.(js|jsx|ts|tsx)'
-  ],
-  setupFiles: [
-    '<rootDir>/tests/unit/setup.js'
   ],
   testURL: 'http://localhost/',
   watchPlugins: [