+++ src/components/child-components/z-knob.vue
@@ -3,17 +3,17 @@
     <svg
       viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg" class="z-range-bar"
       ref="bar"
-      @click.prevent="bar">
+      @click.stop="bar">
         <circle r="52" cx="50" cy="50" :style="[styles]"></circle>
     </svg>
     <svg
       xmlns="http://www.w3.org/2000/svg" class="z-range-bar-bar"
       :style="circleStyle"
       @touchstart="drag = true"
-      @touchmove.prevent="handleBar"
+      @touchmove.stop="handleBar"
       @touchend="drag = false"
       @mousedown="drag = true"
-      @mousemove.prevent="handleBar"
+      @mousemove.stop="handleBar"
       @mouseup="drag = false">
         <circle r="8" cx="20" cy="20" class="z-range-bar-handlebar"></circle>
     </svg>
@@ -47,7 +47,7 @@ export default {
   },
   computed: {
     position () {
-      var diameter = this.$zircle.getComponentWidth(this.$parent.size) / 2
+      const diameter = this.$zircle.getComponentWidth(this.$parent.size) / 2
       return {
         X: (diameter - 3) * Math.cos(this.angle * (Math.PI / 180)),
         Y: (diameter - 3) * Math.sin(this.angle * (Math.PI / 180)),
@@ -55,7 +55,7 @@ export default {
       }
     },
     styles () {
-      var circleLength = 2 * Math.PI * 50
+      const circleLength = 2 * Math.PI * 50
       return {
         transformOrigin: '50% 50%',
         transform: 'rotate(0deg)',
@@ -72,27 +72,23 @@ export default {
     }
   },
   watch: {
-    qty () {
-      this.angle = Math.round(((this.qty - this.min) * 360) / (this.max - this.min))
-    }
   },
   methods: {
     bar (e) {
       e = e.changedTouches ? e.changedTouches[0] : e
       const dimensions = this.$refs.bar.getBoundingClientRect()
-      var centerx = (dimensions.width / 2) + dimensions.left
-      var centery = (dimensions.height / 2) + dimensions.top
-      var posx = e.pageX
-      var posy = e.pageY
-      var deltay = centery - posy
-      var deltax = centerx - posx
-      var tangle = Math.atan2(deltay, deltax) * (180 / Math.PI)
+      const centerx = (dimensions.width / 2) + dimensions.left
+      const centery = (dimensions.height / 2) + dimensions.top
+      const posx = e.pageX
+      const posy = e.pageY
+      const deltay = centery - posy
+      const deltax = centerx - posx
+      let tangle = Math.atan2(deltay, deltax) * (180 / Math.PI)
       tangle -= 180
       tangle = Math.round(tangle)
       if (tangle < 0) tangle = 360 + tangle
-      var prevAngle = Math.round(this.angle)
-      var vm = this
-      var id = setInterval(function () {
+      let prevAngle = Math.round(this.angle)
+      const id = setInterval(() => {
         if (prevAngle > tangle) {
           prevAngle--
         } else if (prevAngle < tangle) {
@@ -100,21 +96,21 @@ export default {
         } else {
           clearInterval(id)
         }
-        vm.angle = prevAngle
-        vm.$emit('update:qty', Math.round((prevAngle / 360) * (vm.max - vm.min)) + vm.min)
+        this.angle = prevAngle
+        this.$emit('update:qty', Math.round((prevAngle / 360) * (this.max - this.min)) + this.min)
       }, 0)
     },
     handleBar (e) {
       if (this.drag === true) {
         e = e.changedTouches ? e.changedTouches[0] : e
         const dimensions = this.$refs.bar.getBoundingClientRect()
-        var centerx = (dimensions.width / 2) + dimensions.left
-        var centery = (dimensions.height / 2) + dimensions.top
-        var posx = e.pageX
-        var posy = e.pageY
-        var deltay = centery - posy
-        var deltax = centerx - posx
-        var tangle = Math.atan2(deltay, deltax) * (180 / Math.PI)
+        const centerx = (dimensions.width / 2) + dimensions.left
+        const centery = (dimensions.height / 2) + dimensions.top
+        const posx = e.pageX
+        const posy = e.pageY
+        const deltay = centery - posy
+        const deltax = centerx - posx
+        let tangle = Math.atan2(deltay, deltax) * (180 / Math.PI)
         tangle -= 180
         tangle = Math.round(tangle)
         if (tangle < 0) tangle = 360 + tangle
@@ -125,7 +121,6 @@ export default {
   },
   mounted () {
     this.angle = Math.round(((this.qty - this.min) * 360) / (this.max - this.min))
-    this.$emit('update:qty', this.qty)
   }
 }
 </script>