+++ src/components/child-components/z-slider.vue
@@ -15,26 +15,26 @@ export default {
   },
   computed: {
     styles () {
-      var zwidth = this.$parent.size
+      const zwidth = this.$parent.size
+      let strokeWidth = 8
       if (zwidth === 'xxl') {
-        var strokeWidth = 3
+        strokeWidth = 3
       } else if (zwidth === 'large' || zwidth === 'xl') {
         strokeWidth = 7
-      } else if (zwidth === 'medium') {
-        strokeWidth = 8
       } else if (zwidth === 'small') {
         strokeWidth = 9
       } else if (zwidth === 'xs' || zwidth === 'extrasmall') {
         strokeWidth = 10
       } else if (zwidth === 'xxs') {
+        strokeWidth = 11
       }
-      var circleLength = 2 * Math.PI * 51
+      const circleLength = 2 * Math.PI * 51
       return {
         transformOrigin: '50% 50%',
         transform: 'rotate(-90deg)',
         strokeDasharray: circleLength,
         strokeDashoffset: circleLength - (this.progress * circleLength / 100),
-        strokeWidth: strokeWidth
+        strokeWidth
       }
     }
   }