+++ src/components/z-canvas.vue
@@ -2,7 +2,7 @@
   <div
     id="z-container"
     class="z-canvas"
-    :class="[classes, $zircle.getTheme(), $zircle.getThemeMode()]"
+    :class="[classes, $zircle.getTheme(), $zircle.getThemeMode(), $zircle.getThemeShape()]"
     :style="[$zircle.getPreviousViewName() !== '' ? {cursor: 'zoom-out'} : {}]"
     @click.stop="goback">
       <div id="z-zoomable-layer" ref="zoom" :style="zoom" @transitionend="notify">
@@ -26,8 +26,7 @@ export default {
   },
   computed: {
     zoom () {
-      var pos = {}
-      this.$zircle.getHistoryLength() === 0 ? pos = { X: 0, Y: 0, Xi: 0, Yi: 0, scale: 1, scalei: 1 } : pos = this.$zircle.getCurrentPosition()
+      const pos = this.$zircle.getHistoryLength() === 0 ? { X: 0, Y: 0, Xi: 0, Yi: 0, scale: 1, scalei: 1 } : this.$zircle.getCurrentPosition()
       return {
         transform: 'scale(' + pos.scale + ') translate3d(' + pos.Xi + 'px, ' + pos.Yi + 'px, 0px)',
         transition: 'transform 1000ms ease-in-out'
@@ -57,21 +56,9 @@ export default {
     this.$zircle.setComponentList(this.views)
   },
   mounted () {
-    var vm = this
     // Get window dimension to set the initial width of ui components such as z-panel
-    this.$nextTick()
-      .then(function () {
-      // DOM updated
-        vm.$zircle.getDimensions()
-      })
-    window.addEventListener('resize', function (event) {
-      // On resize change the width of ui components
-      vm.$zircle.getDimensions()
-    })
-    document.onmouseleave = function () {
-      // User's mouse has left the page.
-      vm.$zircle.setNavigationMode('backward')
-    }
+    this.$nextTick().then(() => this.$zircle.getDimensions())
+    window.addEventListener('resize', () => this.$zircle.getDimensions())
   }
 }
 </script>