+++ src/components/z-list.vue
@@ -3,28 +3,32 @@
     <div v-for="(item, index) in $zircle.getCurrentPage()" :key="item[0] + '-' + index">
       <slot v-bind="item" :index=index></slot>
     </div>
-    <z-pagination  v-if="$zircle.getNumberOfPages() > 1"
-      v-for="(page, index) in $zircle.getNumberOfPages()"
-      :key="index + '_page'"
-      :index="index"
-      :distance="112"
-      :angle="(180 - (180 - ($zircle.getNumberOfPages() * 10))) / $zircle.getNumberOfPages() * ($zircle.getNumberOfPages() - index) + ((180 - (180 - (180 - ($zircle.getNumberOfPages() * 10)))) - ((180 - (180 - ($zircle.getNumberOfPages() * 10))) / $zircle.getNumberOfPages())) / 2"
-      :active="$zircle.getCurrentPageIndex()"
-      @mouseover.native = "$zircle.allowBackwardNavigation(true)"
-      @mouseleave.native = "$zircle.allowBackwardNavigation(false)"
-      @click.native="$zircle.setCurrentPageIndex(index)"/>
+    <template v-if="$zircle.getThemeShape() === 'circle' && $zircle.getNumberOfPages() > 1">
+      <z-pagination
+        v-for="(page, index) in $zircle.getNumberOfPages()"
+        :key="index + '_page'"
+        :index="index"
+        :distance="$zircle.getComponentWidth(size) + $zircle.getComponentWidth('xs') + 10"
+        :angle="(180 - (180 - ($zircle.getNumberOfPages() * 10))) / $zircle.getNumberOfPages() * ($zircle.getNumberOfPages() - index) + ((180 - (180 - (180 - ($zircle.getNumberOfPages() * 10)))) - ((180 - (180 - ($zircle.getNumberOfPages() * 10))) / $zircle.getNumberOfPages())) / 2"
+        :active="$zircle.getCurrentPageIndex()"
+        @mouseover="$zircle.allowBackwardNavigation(true)"
+        @mouseleave="$zircle.allowBackwardNavigation(false)"
+        @click.native="$zircle.setCurrentPageIndex(index)"/>
+    </template>
   </section>
 </template>
 
 <script>
 import ZPagination from './child-components/z-pagination'
+
 function chunk (myArray, chunkSize) {
-  var res = []
+  const res = []
   while (myArray.length) {
     res.push(myArray.splice(0, chunkSize))
   }
   return res
 }
+
 export default {
   name: 'z-list',
   props: {
@@ -35,6 +39,10 @@ export default {
     items: {
       type: Array,
       required: true
+    },
+    square: {
+      type: Boolean,
+      default: false
     },
     perPage: {
       type: [Number],