+++ src/demo/home.vue
@@ -1,23 +1,24 @@
 <template>
-  <z-view style="border-width: 8px;">
-    WELCOME TO ZIRCLE UI
-    <div slot='extension'>
+  <z-view size="xl" square>
+    <template #extension>
       <z-spot
+        size="xs"
         :angle='45'
+        :distance='130'
         to-view='docs'
-        class='accent'
-        label='Docs'>
-        <i class='fas fa-book'></i>
+        class='shade'
+        label='doc & resources'>
+        <i class="fas fa-question-circle"></i>
       </z-spot>
-      <z-spot
-        button
-        :angle='135'
-        class='accent'
-        label='Github'
-        @click.native="openUrl">
-        <i class='fab fa-github'></i>
+       <z-spot
+        size="m"
+        :angle='0'
+        :distance='0'
+        class="shade"
+        to-view='demo'>
+        <b>Zoom me</b>
       </z-spot>
-    </div>
+    </template>
   </z-view>
 </template>
 <script>